requirements.txt
slack_bolt==1.18.0
google-generativeai==0.3.2
python-dotenv==1.0.0
psycopg2-binary>=2.9.10
SQLAlchemy>=2.0.36
alembic>=1.13.2
gunicorn>=22.0.0
google-api-python-client==2.111.0
google-auth==2.26.1
asana==3.2.2
pandas>=2.2.3
reportlab==4.0.9
pdfplumber==0.10.3
python-docx==1.1.0
apscheduler==3.10.4
